requirements.txt
numpy==1.21.5
opencv-python-headless
Pillow==9.4.0
spacy
torchvision
jsonlines==3.1.0

requests~=2.26.0
setuptools~=58.0.4
